First compiling version of WIN32 port of SMSLib and game engine

+++ project/source/engine/commands/background_commands.c
@@ -5,5 +5,5 @@
 
 void SetBackgroundVerticalScroll(s16 value)
 {
-	SMS_setBGScrollY(value);
+	SMS_setBGScrollY((u8)value);
 }
